+++ Sources/APRelay/App/QueueName+APRelay.swift
@@ -1,0 +1,6 @@
+import Queues
+
+extension QueueName {
+    /// The queue used for delivering activities to subscriber inboxes.
+    static let delivery = QueueName(string: "delivery")
+}